# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      tasks_average_completion_time:
        example: 1m30s
        type: string
      tasks_completed:
        example: 10
        type: integer
      tasks_completed_rate:
        example: 20
        type: number
      tasks_created:
        example: 50
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        type: boolean
      description:
        example: "null"
        type: string
      title:
        example: Погулять с собакой
        type: string
    type: object
  internal_features_tasks_transport_http.PatchUserResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Сделать до четверга домашнее задание по математике
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Домашка
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+799888899"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+799988877"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+799988877"
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Максим Максимович
        type: string
      phone_number:
        example: "+7888111231"
        type: string
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+799988877"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
  /statistics:
    get:
      description: Получение статистики по задачам с опциональной фильтрацией по user_id
        и/или временному промежутку
      parameters:
      - description: Фильтрация статистики по конкретному пользователю
        format: uuid
        in: query
        name: user_id
        type: string
      - description: 'Начало промежутка рассмотрения статистики (включительно), формат:
          YYYY-MM-DD'
        in: query
        name: from
        type: string
      - description: 'Конец промежутся рассмотрения статистики (не включительно),
          формат: YYYY-MM-DD'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение статистики
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение статистики
      tags:
      - statistics
  /tasks:
    get:
      description: Просмотр списка задач с опциональной пагинацией и/или фильтрацией
        по ID автора задачи
      parameters:
      - description: Фильтрация задач по ID автора
        format: uuid
        in: query
        name: user_id
        type: string
      - description: Размер страницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в системе
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удалить существующую в системе задачу по её ID
      parameters:
      - description: ID удаляемой задачи
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Успешное удаление задачи
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Удаление задачи
      tags:
      - tasks
    get:
      description: Получение конкретной задачи по её ID
      parameters:
      - description: ID получаемой задачи
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешной найдена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задачи
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Обновляет информацию об уже существующей в системе задаче
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: `description` игнорируется, значение в БД не меняется
        2. **Явно передано значение**: `"description": "Утром в 06:30 выйти на прогулку с Бобиком"` — устанавливает новое описание для задачи
        3. **Явно передан null**: `"description": null`— очищает поле в БД (set to NULL)
        Ограничения: `title` и `completed` не могут быть выставлены как null
      parameters:
      - description: ID изменяемой задачи
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
  /users:
    get:
      description: Получение списка пользователей
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаление существующего пользователя в системе
      parameters:
      - description: ID удаляемого пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление  пользователя
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Получение конкретного пользователя в системе по его id
      parameters:
      - description: ID получаемого пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно найден
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Изменения информации об уже существующем в системе пользователе
      parameters:
      - description: ID изменяемого пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: успешно измененный пользователь
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_ornstein77_Golang-TodoApp_internal_core_transport_http_response.ErrorResponse'
      summary: Изменение пользователя
      tags:
      - users
swagger: "2.0"