# internal/features/tasks/transport/http/delete_task.go
package tasks_transport_http

import (
	"net/http"

	core_logger "github.com/ornstein77/Golang-TodoApp/internal/core/logger"
	core_http_request "github.com/ornstein77/Golang-TodoApp/internal/core/transport/http/request"
	core_http_response "github.com/ornstein77/Golang-TodoApp/internal/core/transport/http/response"
)

// DeleteTask    godoc
// @Summary      Удаление задачи
// @Description  Удалить существующую в системе задачу по её ID
// @Tags         tasks
// @Param        id   path  string  true "ID удаляемой задачи" Format(uuid)
// @Success      204  "Успешное удаление задачи"
// @Failure      400  {object}  core_http_response.ErrorResponse "Bad request"
// @Failure      404  {object}  core_http_response.ErrorResponse "Task not found"
// @Failure      500  {object}  core_http_response.ErrorResponse "Internal server error"
// @Router       /tasks/{id} [delete]
func (h *TasksHTTPHandler) DeleteTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	taskID, err := core_http_request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to get ttaskID path value",
		)
		return
	}
	if err := h.tasksService.DeleteTask(ctx, taskID); err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to delete task",
		)
		return
	}
}